Editor: Add a featured image dropzone above the post title.
This restores functionality that existed in the old editor.

The existing `MediaLibraryDropZone` is adjusted to make room for the new dropzone, because the `MediaLibraryDropZone` previously took up the entire editor pane.

The `line-height` was removed from `.editor-featured-image` because it was causing the text and icon to be squished, and didn't seem to be needed anywhere else in Calypso.

Fixes #13277