Load multiple pages of posts sequentially if list is scrolled to end
This also (eventually) resolves an issue where the placeholder at the
end of the list would never disappear if the list content was loaded
from persisted state.

After this commit, the placeholder will stick around until each page of
the posts query is repeated, then disappear.  We should arguably avoid
re-fetching pages of cached content, but currently there is a
disconnect:  we store the results of the post fetching operations, but
not the page numbers that were fetched.

Finally, this commit also simplifies the logic for displaying the
placeholder at the end of the posts list.